Rename TermTable -> TermTableRow
These are going to be representing the individual rows, as there wasn't
any logic relating to the tables in these classes.
