+++ lib/term_table_row/base.rb
@@ -1,4 +1,4 @@
-module TermTable
+module TermTableRow
   class Base < Scraped::HTML
     def self.expected_headers(expected = nil)
       @expected_headers ||= expected
